requirements.txt
pyrogram
tgcrypto
uvloop
wget
python-dotenv